refactor(desktop): isolate project context state
